Fix Chat circular deps
